implement correct MVC structure for GameViewController (Controller) and GameScene (View)
