+++ DartMania/Protocols/EndGameDecisionDelegate.swift
@@ -1,0 +1,14 @@
+//
+//  EndGameDecisionDelegate.swift
+//  DartMania
+//
+//  Created by Tim Hehmann on 04.03.19.
+//  Copyright © 2019 Tim Hehmann. All rights reserved.
+//
+
+import Foundation
+
+protocol EndGameDecisionDelegate : class {
+    func didTapReturnToMenuButton()
+    func didTapRestartButton()
+}
