+++ DartMania/ViewController/GameSettingsViewController.swift
@@ -8,10 +8,6 @@
 
 import UIKit
 
-protocol GameSettingsSelectionDelegate: class {
-    func didPickGameSettings(settings: DartGameSettings)
-}
-
 class GameSettingsViewController: UIViewController {
     
     weak var gameSettingsSelectionDelegate: GameSettingsSelectionDelegate!
@@ -94,17 +90,6 @@ class GameSettingsViewController: UIViewController {
         gameSettingsSelectionDelegate.didPickGameSettings(settings: self.settings)
         dismiss(animated: true, completion: nil)
     }
-    
-
-    /*
-    // MARK: - Navigation
-
-    // In a storyboard-based application, you will often want to do a little preparation before navigation
-    override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
-        // Get the new view controller using segue.destination.
-        // Pass the selected object to the new view controller.
-    }
-    */
     
     override var prefersStatusBarHidden: Bool {
         return true
